# use 32 bit index to improve expand op